Extract SupersetClient setup into a separate file and fix translation link
