+++ superset/assets/src/common.js
@@ -1,9 +1,13 @@
-/* eslint global-require: 0, no-console: 0 */
+/* eslint global-require: 0 */
 import $ from 'jquery';
 import 'abortcontroller-polyfill/dist/abortcontroller-polyfill-only';
 import { SupersetClient } from '@superset-ui/core';
 import { toggleCheckbox } from './modules/utils';
 import setupColors from './setup/setupColors';
+import setupClient from './setup/setupClient';
+
+setupClient();
+setupColors();
 
 $(document).ready(function () {
   $(':checkbox[data-checkbox-api-prefix]').change(function () {
@@ -16,15 +20,15 @@ $(document).ready(function () {
   // for language picker dropdown
   $('#language-picker a').click(function (ev) {
     ev.preventDefault();
-
-    SupersetClient.get({ endpoint: ev.currentTarget.href })
+    SupersetClient.get({
+      endpoint: ev.currentTarget.getAttribute('href'),
+      parseMethod: null,
+    })
       .then(() => {
         location.reload();
       });
   });
 });
-
-setupColors();
 
 export function appSetup() {
   // A set of hacks to allow apps to run within a FAB template
@@ -32,17 +36,4 @@ export function appSetup() {
   window.$ = $;
   window.jQuery = $;
   require('bootstrap');
-
-  const csrfNode = document.querySelector('#csrf_token');
-  const csrfToken = csrfNode ? csrfNode.value : null;
-
-  SupersetClient.configure({
-    protocol: (window.location && window.location.protocol) || '',
-    host: (window.location && window.location.host) || '',
-    csrfToken,
-  })
-    .init()
-    .catch((error) => {
-      console.warn('Error initializing SupersetClient', error);
-    });
 }
